refactor: replace all uint32 to uint256

+++ contracts/AppStorage.sol
@@ -15,7 +15,7 @@ struct AppStorage {
     mapping(bytes32 => bytes) versionNumbers; // keccak(name,branch) => <bytes3[]> versionNumbers
     mapping(bytes32 => string[]) branches; // keccak(name) => string[]
     mapping(bytes32 => VersionInfo) versions; // keccak(name,branch,major,minor,patch) => VersionInfo>
-    mapping(bytes32 => uint32) moduleIdxs; // key - keccak256(name) => value - index of element in "s.modules" array
+    mapping(bytes32 => uint256) moduleIdxs; // key - keccak256(name) => value - index of element in "s.modules" array
     mapping(bytes32 => EnumerableSet.UintSet) modsByContextType; // key - keccak256(contextId, owner), value - index of element in "s.modules" array
     mapping(bytes32 => EnumerableSet.AddressSet) adminsOfModules; // key - mod_name => EnumerableSet address for added, removed and get all address
     mapping(bytes32 => EnumerableStringSet.StringSet) contextIdsOfModules; // key - mod_name => EnumerableSet
